Start on the electron-redux refactor. Next up: - connect main and render store - sync them - make it possible to save the store - clean up the build script a bit more
