--- app/index.js
@@ -1,55 +1,0 @@
-import React from 'react';
-import { render } from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
-import injectTapEventPlugin from 'react-tap-event-plugin';
-
-// import Root from './containers/Root';
-import { configureStore, history } from './store/configureStore';
-import './app.global.css';
-
-import App from './components/refactor/App';
-
-// Needed for onTouchTap
-// http://stackoverflow.com/a/34015469/988941
-injectTapEventPlugin();
-
-const store = configureStore();
-
-/* render(
-  <AppContainer>
-    <Root store={store} history={history} />
-  </AppContainer>,
-  document.getElementById('root')
-);
-
-if (module.hot) {
-  module.hot.accept('./containers/Root', () => {
-    const NextRoot = require('./containers/Root'); // eslint-disable-line global-require
-    render(
-      <AppContainer>
-        <NextRoot store={store} history={history} />
-      </AppContainer>,
-      document.getElementById('root')
-    );
-  });
-}*/
-
-
-render(
-  <AppContainer>
-    <App store={store} history={history} />
-  </AppContainer>,
-  document.getElementById('root')
-);
-
-if (module.hot) {
-  module.hot.accept('./components/refactor/App', () => {
-    const NextApp = require('./components/refactor/App'); // eslint-disable-line global-require
-    render(
-      <AppContainer>
-        <NextApp store={store} history={history} />
-      </AppContainer>,
-      document.getElementById('root')
-    );
-  });
-}
