+++ app/package.json
@@ -1,9 +1,9 @@
 {
   "name": "moviecast-desktop",
   "productName": "MovieCast",
-  "version": "1.0.0",
-  "description": "Electron application boilerplate based on React, React Router, Webpack, React Hot Loader for rapid application development",
-  "main": "./main.prod.js",
+  "version": "0.0.1",
+  "description": "Watch movies like never before.",
+  "main": "./main/main.prod.js",
   "author": {
     "name": "Dizzy",
     "email": "itsdizzy40@gmail.com",
