+++ app/renderer/app.html
@@ -24,7 +24,7 @@
         // Dynamically insert the DLL script in development env in the
         // renderer process
         if (process.env.NODE_ENV === 'development') {
-          scripts.push('../dll/vendor.dll.js');
+          scripts.push('../../dll/vendor.dll.js');
         }
 
         // Dynamically insert the bundled app script in the renderer process
