+++ app/renderer/components/Catalog.jsx
