+++ app/renderer/components/GridList.js
