+++ app/renderer/components/refactor/App.js
