+++ app/renderer/components/refactor/AppContent.js
