+++ app/renderer/components/refactor/AppFrame.js
