+++ app/renderer/components/refactor/Catalog.js
