+++ app/renderer/containers/App.js
