+++ app/renderer/containers/AppToolbar.js
