+++ app/renderer/containers/MoviesCatalog.js
