+++ app/renderer/index.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import { render } from 'react-dom';
+import { AppContainer } from 'react-hot-loader';
+
+import { configureStore, history } from '../shared/store/configureStore';
+import './app.global.css';
+
+import App from './components/refactor/App';
+
+const store = configureStore();
+
+render(
+  <AppContainer>
+    <App store={store} history={history} />
+  </AppContainer>,
+  document.getElementById('root')
+);
+
+if (module.hot) {
+  module.hot.accept('./components/refactor/App', () => {
+    const NextApp = require('./components/refactor/App'); // eslint-disable-line global-require
+    render(
+      <AppContainer>
+        <NextApp store={store} history={history} />
+      </AppContainer>,
+      document.getElementById('root')
+    );
+  });
+}
