+++ app/shared/actions/settings.js
