+++ app/shared/reducers/electron.js
