+++ app/shared/reducers/index.js
