+++ app/shared/reducers/settings.js
