+++ webpack.config.renderer.dev.js
@@ -40,7 +40,7 @@ export default merge.smart(baseConfig, {
     'react-hot-loader/patch',
     `webpack-dev-server/client?http://localhost:${port}/`,
     'webpack/hot/only-dev-server',
-    path.join(__dirname, 'app/index.js'),
+    path.join(__dirname, 'app/renderer/index.js'),
   ],
 
   output: {
